fix(diag): /proc/temp — 64-bit time_t crash + working XADC die temp
Two real bugs surfaced getting xadc_int reading true:

1. Data-abort in the fs task on `cat /proc/temp`: `_gettimeofday` fills a
   struct timeval whose tv_sec is 64-bit here, but four call sites declared
   `struct { long sec, usec; }` (8 bytes) — tv_usec spilled past the struct onto
   the adjacent stack local (the procfs buffer pointer), and the next write went
   through garbage.  Latent in pf_gen_uptime / pf_gen_video / devfs random-seed
   too; all widened to `long long`.  (DFAR was always a µs-range value — the tell.)

2. XADC: enabling the interface isn't enough (needs reset released + arbiter on),
   the SPI-like FIFO must never be popped empty (that SLVERRs -> external abort),
   and the read must settle the command pipeline (read the reg twice, pop 3).
   The result comes back at exactly half scale on this path — verified on HW:
   raw*2 gives the die temp AND correct Vccint 1.0V / Vccaux 1.8V, with all
   config regs zero (not bipolar, not averaging) — so scale up one bit.

Reads ~76C die vs ~50C board.

+++ loader/test/freertos/vfs_devfs.c
@@ -225,7 +225,7 @@ static int dv_open(vfs_mount *m, const char *rel, int flags, vfs_file *f)
         return 0;
     }
     if (d->rd == dv_rand_rd) {           /* per-open xorshift state, clock-seeded */
-        struct { long sec, usec; } tv = { 0, 0 };
+        struct { long long sec, usec; } tv = { 0, 0 };   /* time_t is 64-bit here — must not undersize */
         _gettimeofday(&tv, 0);
         uint32_t s = (uint32_t)tv.usec ^ ((uint32_t)tv.sec << 12) ^ 0x9e3779b9u;
         f->priv = (void *)(uintptr_t)(s ? s : 1);
